backend/rag/requirements.txt
watchdog
pdfminer.six
google-genai
python-dotenv